Normalized
Still with probability of recovery; added a counter for nodes that have no connection and subtract that number from the total number of nodes and use this new number to calculate the fraction of healthy, sick, etc. when writing to the .csv file.  So now its scaled to the nodes that have connections.